- 메인 액티비티 수정
